pyproject.toml
[tool.poetry]
name = "jobby"
version = "1.0.0"
description = "API for resume parsing and candidate evaluation"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "1.0.0"
SQLAlchemy = "^2.0.38"
PyMySQL = "1.1.1"
python-crontab = "3.0.0"
python-docx = "1.1.2"
pdfminer-six = "20240706"

aiohttp = "3.9.1"
asyncio = "3.4.3"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
openai = "^1.0.0"
boto3 = "^1.36.23"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
pytest-asyncio = "0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = 'app:app'